Move this logic to a TestRule implementation and hook it up to SolrTestCase.
