+++ solr/core/src/test/org/apache/solr/search/TestRangeQuery.java
@@ -43,6 +43,7 @@ import org.apache.solr.response.SolrQueryResponse;
 import org.apache.solr.schema.FieldType;
 import org.apache.solr.schema.NumberType;
 import org.apache.solr.schema.StrField;
+import org.apache.solr.util.QueryLimitsTestInjectionRule;
 import org.apache.solr.util.TestInjection;
 import org.junit.After;
 import org.junit.Before;
@@ -463,7 +464,7 @@ public class TestRangeQuery extends SolrTestCaseJ4 {
             queryService.awaitTermination(
                 1, TimeUnit.SECONDS)); // All queries after should be very fast
 
-        if (withQueryTimeout) {
+        if (QueryLimitsTestInjectionRule.enabled) {
           assertTrue(
               "Create multiple DocSet-s outside of cache because of possible query timeouts",
               TestInjection.countDocSetDelays.get() > 0);
